add manuscript resubmission rules to frontend
