+++ apps/crn-server/test/fixtures/teams.fixtures.ts
@@ -236,6 +236,7 @@ export const getTeamDataObject = (): TeamDataObject => ({
           lifecycle: 'Preprint',
           type: 'Original Research',
           description: 'A good description',
+          count: 1,
           createdBy: manuscriptAuthor,
           updatedBy: manuscriptAuthor,
           createdDate: '2020-09-23T20:45:22.000Z',
@@ -273,6 +274,7 @@ export const getTeamDataObject = (): TeamDataObject => ({
           lifecycle: 'Preprint',
           type: 'Original Research',
           description: 'A good description',
+          count: 1,
           createdBy: manuscriptAuthor,
           updatedBy: manuscriptAuthor,
           createdDate: '2020-09-23T20:45:22.000Z',
