+++ packages/contentful/src/crn/autogenerated-gql/gql.ts
@@ -79,7 +79,7 @@ const documents = {
     types.FetchInterestGroupsByUserIdDocument,
   '\n  query FetchLabs($limit: Int, $skip: Int, $where: LabsFilter) {\n    labsCollection(limit: $limit, skip: $skip, where: $where, order: name_ASC) {\n      total\n      items {\n        sys {\n          id\n        }\n        name\n      }\n    }\n  }\n':
     types.FetchLabsDocument,
-  '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n':
+  '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_firstPublishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        count\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n':
     types.ManuscriptsContentFragmentDoc,
   '\n  query FetchManuscriptById($id: String!, $fetchReplies: Boolean = false) {\n    manuscripts(id: $id) {\n      ...ManuscriptsContent\n      teamsCollection(limit: 10) {\n        items {\n          sys {\n            id\n          }\n        }\n      }\n    }\n  }\n  \n':
     types.FetchManuscriptByIdDocument,
@@ -359,8 +359,8 @@ export function gql(
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function gql(
-  source: '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n',
-): (typeof documents)['\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_publishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n'];
+  source: '\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_firstPublishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        count\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n',
+): (typeof documents)['\n  fragment ManuscriptsContent on Manuscripts {\n    sys {\n      id\n    }\n    title\n    status\n    count\n    versionsCollection(limit: 20, order: sys_firstPublishedAt_DESC) {\n      items {\n        sys {\n          id\n          publishedAt\n          firstPublishedAt\n        }\n        type\n        lifecycle\n        description\n        count\n        manuscriptFile {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        keyResourceTable {\n          sys {\n            id\n          }\n          fileName\n          url\n        }\n        additionalFilesCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            fileName\n            url\n          }\n        }\n        preprintDoi\n        publicationDoi\n        requestingApcCoverage\n        submitterName\n        submissionDate\n        otherDetails\n        acknowledgedGrantNumber\n        acknowledgedGrantNumberDetails {\n          ...DiscussionsContent\n        }\n        asapAffiliationIncluded\n        asapAffiliationIncludedDetails {\n          ...DiscussionsContent\n        }\n        manuscriptLicense\n        manuscriptLicenseDetails {\n          ...DiscussionsContent\n        }\n        datasetsDeposited\n        datasetsDepositedDetails {\n          ...DiscussionsContent\n        }\n        codeDeposited\n        codeDepositedDetails {\n          ...DiscussionsContent\n        }\n        protocolsDeposited\n        protocolsDepositedDetails {\n          ...DiscussionsContent\n        }\n        labMaterialsRegistered\n        labMaterialsRegisteredDetails {\n          ...DiscussionsContent\n        }\n        availabilityStatement\n        availabilityStatementDetails {\n          ...DiscussionsContent\n        }\n        teamsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            displayName\n            inactiveSince\n          }\n        }\n        labsCollection(limit: 10) {\n          items {\n            sys {\n              id\n            }\n            name\n          }\n        }\n        createdBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        updatedBy {\n          sys {\n            id\n          }\n          firstName\n          nickname\n          lastName\n          alumniSinceDate\n          avatar {\n            url\n          }\n          teamsCollection(limit: 3) {\n            items {\n              team {\n                sys {\n                  id\n                }\n                displayName\n              }\n            }\n          }\n        }\n        firstAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        additionalAuthorsCollection(limit: 15) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        correspondingAuthorCollection(limit: 1) {\n          items {\n            __typename\n            ... on ExternalAuthors {\n              sys {\n                id\n              }\n              name\n              email\n            }\n            ... on Users {\n              sys {\n                id\n              }\n              avatar {\n                url\n              }\n              firstName\n              lastName\n              nickname\n              email\n            }\n          }\n        }\n        linkedFrom {\n          complianceReportsCollection(limit: 1) {\n            items {\n              url\n              description\n              count\n            }\n          }\n        }\n      }\n    }\n  }\n  \n'];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
