+++ packages/react-components/src/organisms/__tests__/ManuscriptVersionCard.test.tsx
@@ -1,13 +1,8 @@
 import {
   createDiscussionResponse,
   createManuscriptResponse,
-  createUserResponse,
 } from '@asap-hub/fixtures';
-import {
-  ManuscriptLifecycle,
-  ManuscriptVersion,
-  UserTeam,
-} from '@asap-hub/model';
+import { ManuscriptLifecycle, ManuscriptVersion } from '@asap-hub/model';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { createMemoryHistory } from 'history';
@@ -16,8 +11,6 @@ import { Router } from 'react-router-dom';
 import ManuscriptVersionCard, {
   getLifecycleCode,
   getManuscriptVersionUID,
-  isManuscriptAuthor,
-  isManuscriptLead,
 } from '../ManuscriptVersionCard';
 
 const setScrollHeightMock = (height: number) => {
@@ -38,7 +31,6 @@ afterAll(jest.clearAllMocks);
 
 const baseVersion = createManuscriptResponse().versions[0] as ManuscriptVersion;
 const props: ComponentProps<typeof ManuscriptVersionCard> = {
-  user: null,
   version: baseVersion,
   grantId: '000123',
   teamIdCode: 'TI1',
@@ -48,6 +40,7 @@ const props: ComponentProps<typeof ManuscriptVersionCard> = {
   getDiscussion: jest.fn(),
   manuscriptId: 'manuscript-1',
   isTeamMember: true,
+  canEditManuscript: true,
 };
 
 it('displays quick checks when present', () => {
@@ -209,21 +202,9 @@ it('displays createdBy as fallback for updatedBy when updatedBy is well defined'
 });
 
 describe('edit', () => {
-  const baseUser = createUserResponse({}, 1);
-  const user = {
-    ...baseUser,
-    teams: [
-      {
-        ...baseUser.teams[0],
-        id: 'team-1',
-        role: 'Project Manager',
-      },
-    ] as UserTeam[],
-    algoliaApiKey: 'algolia-mock-key',
-  };
-  it('does not display the edit button when no user is set', () => {
+  it('does not display the edit button when canEditManuscript is false', () => {
     const { queryByLabelText } = render(
-      <ManuscriptVersionCard {...props} user={null} />,
+      <ManuscriptVersionCard {...props} canEditManuscript={false} />,
     );
     expect(queryByLabelText('Edit')).not.toBeInTheDocument();
   });
@@ -234,72 +215,13 @@ describe('edit', () => {
 
     const { getByLabelText } = render(
       <Router history={history}>
-        <ManuscriptVersionCard {...props} user={user} />
+        <ManuscriptVersionCard {...props} />
       </Router>,
     );
     userEvent.click(getByLabelText('Edit'));
     expect(pushSpy).toHaveBeenCalledWith(
       '/network/teams/team-id-0/workspace/edit-manuscript/manuscript-1',
     );
-  });
-
-  describe('isManuscriptAuthor', () => {
-    it('returns true when user is author', () => {
-      expect(
-        isManuscriptAuthor({
-          authors: [{ ...createUserResponse(), id: 'user-test' }],
-          user: {
-            ...user,
-            id: 'user-test',
-          },
-        }),
-      ).toBeTruthy();
-    });
-    it('returns false when user is not author', () => {
-      expect(
-        isManuscriptAuthor({
-          authors: [{ ...createUserResponse(), id: 'different-user' }],
-          user: {
-            ...user,
-          },
-        }),
-      ).not.toBeTruthy();
-    });
-  });
-
-  describe('isManuscriptLead', () => {
-    it('returns true when user is team lead', () => {
-      expect(
-        isManuscriptLead({
-          version: baseVersion,
-          user: {
-            ...user,
-            teams: [
-              {
-                id: 'team-1',
-                role: 'Project Manager',
-              },
-            ],
-          },
-        }),
-      ).toBeTruthy();
-    });
-    it('returns false when user is not team lead', () => {
-      expect(
-        isManuscriptLead({
-          version: baseVersion,
-          user: {
-            ...user,
-            teams: [
-              {
-                id: 'team-1',
-                role: 'Scientific Advisory Board',
-              },
-            ],
-          },
-        }),
-      ).not.toBeTruthy();
-    });
   });
 });
