+++ packages/react-components/src/templates/ManuscriptForm.tsx
@@ -90,6 +90,7 @@ type OptionalVersionFields = Array<
     | 'type'
     | 'lifecycle'
     | 'complianceReport'
+    | 'count'
     | 'createdBy'
     | 'updatedBy'
     | 'createdDate'
@@ -239,6 +240,7 @@ type ManuscriptFormProps = Omit<
   | 'keyResourceTable'
   | 'additionalFiles'
   | 'description'
+  | 'count'
   | 'createdBy'
   | 'updatedBy'
   | 'createdDate'
